Try cmake_system_version for sdk

+++ tests/include/minunit.h
@@ -214,7 +214,7 @@ minunit_suite_teardown = NULL;\
 	minunit_tmp_e = (expected);\
 	minunit_tmp_r = (result);\
 	if (fabs(minunit_tmp_e-minunit_tmp_r) > tol) {\
-		int minunit_significant_figures = 1 - log10(tol);\
+		int minunit_significant_figures = (int) (1 - log10(tol));\
 		snprintf(minunit_last_message, MINUNIT_MESSAGE_LEN, "%s failed:\n\t%s:%d: %.*g expected but was %.*g", __func__, __FILE__, __LINE__, minunit_significant_figures, minunit_tmp_e, minunit_significant_figures, minunit_tmp_r);\
 		minunit_status = 1;\
 		return;\
